+++ cmake/pulp/toolchain_llvm.cmake
@@ -19,6 +19,7 @@ set(CMAKE_EXECUTABLE_SUFFIX ".elf")
 add_compile_options(
   -target riscv32-unknown-elf
   -march=${ISA}
+  -mabi=ilp32f
   -ffunction-sections
   -fdata-sections
   -fomit-frame-pointer
@@ -26,7 +27,7 @@ add_compile_options(
   -O3
   -DNUM_CORES=${NUM_CORES}
   -MP
-  --sysroot=${TOOLCHAIN_INSTALL_DIR}/picolibc/riscv/rv32imc
+  --sysroot=${TOOLCHAIN_INSTALL_DIR}/picolibc/riscv/rv32imf
   -fno-builtin-memcpy
   -fno-builtin-memset
 )
@@ -36,8 +37,9 @@ add_link_options(
   -MP
   -nostartfiles
   -march=${ISA}
-  -L${TOOLCHAIN_INSTALL_DIR}/picolibc/riscv/rv32imc/lib
-  -L${TOOLCHAIN_INSTALL_DIR}/lib/clang/15.0.0/lib/baremetal/rv32imc/
+  -mabi=ilp32f
+  -L${TOOLCHAIN_INSTALL_DIR}/picolibc/riscv/rv32imf/lib
+  -L${TOOLCHAIN_INSTALL_DIR}/lib/clang/15.0.0/lib/baremetal/rv32imf/
   -z norelro
   -fno-builtin-memcpy
   -fno-builtin-memset